proof-of-concept for Abora
It can roundtrip something that sounds a lot
like whistling. The reader produces a large
bitmap, there's a laborious manual
transcription process, and then what comes
out sounds a lot like what came in.

The process could be streamlined a lot, both
in terms of UX and optimisation. Every
possible shortcut was taken.

But the fun part to come is that once we have
a representation of a melody in a computer-
manipulable form, we should be able to process
and synthesise it in more fun ways than just
as a simple sine wave.